+++ src/routes/productRoutes.js
@@ -5,9 +5,7 @@ const productController = require('../controllers/productController');
 const router = express.Router();
 
 router.post('/products', productController.createProduct);
-router.get('/products/:code', productController.getProduct);
-router.put('/products/:code/image', productController.updateProductImage);
-
-// Otras rutas como update, delete, etc.
+router.get('/products/:productId', productController.getProduct);
+router.put('/products/:productId/image', productController.updateProductImage);
 
 module.exports = router;
